Created `gempyor.testing` module
Created a testing module for housing testing utilities starting with
`change_directory_to_temp_directory` pytest fixture. Not sure how to go
about testing this module or if that is necessary.
